+++ app/src/main/java/de/dreier/mytargets/managers/SettingsManager.java
@@ -110,7 +110,7 @@ public class SettingsManager {
 
     public static void setTarget(Target target) {
         preferences.edit()
-                .putInt(KEY_TARGET, (int) target.getId())
+                .putInt(KEY_TARGET, (int)(long)target.getId())
                 .putInt(KEY_SCORING_STYLE, target.scoringStyle)
                 .putInt(KEY_TARGET_DIAMETER_VALUE, target.size.value)
                 .putString(KEY_TARGET_DIAMETER_UNIT, Dimension.Unit.toStringHandleNull(target.size.unit))
